Srinivas branch

+++ hztb-rest-api/src/main/java/com/svs/hztb/api/sm/model/opinion/OpinionResponseInput.java
@@ -1,35 +1,41 @@
 package com.svs.hztb.api.sm.model.opinion;
 
 public class OpinionResponseInput {
-	private int userId;
+	private Long userId;
 	private int opinionReqId;
 	private String responseCode;
 	private String responseTxt;
-	public int getUserId() {
+
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(int userId) {
+
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
+
 	public int getOpinionReqId() {
 		return opinionReqId;
 	}
+
 	public void setOpinionReqId(int opinionReqId) {
 		this.opinionReqId = opinionReqId;
 	}
+
 	public String getResponseCode() {
 		return responseCode;
 	}
+
 	public void setResponseCode(String responseCode) {
 		this.responseCode = responseCode;
 	}
+
 	public String getResponseTxt() {
 		return responseTxt;
 	}
+
 	public void setResponseTxt(String responseTxt) {
 		this.responseTxt = responseTxt;
 	}
-	
-	
 
 }
