+++ hztb-servicemanager/src/main/java/com/svs/hztb/repository/GroupRepository.java
@@ -7,17 +7,16 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
 import com.svs.hztb.entity.GroupEntity;
-import com.svs.hztb.entity.OpinionEntity;
 
 public interface GroupRepository extends JpaRepository<GroupEntity, Integer> {
-//	@Modifying
-//	@Query("UPDATE Space c SET c.owner = :name WHERE c.id = :id")
-//	Integer setNameForId(@Param("name") String name, @Param("id")
+	// @Modifying
+	// @Query("UPDATE Space c SET c.owner = :name WHERE c.id = :id")
+	// Integer setNameForId(@Param("name") String name, @Param("id")
 	@Modifying
 	@Query("update GroupEntity g set g.groupName = ?1 where g.groupId = ?2 and g.groupOwner=?3")
-	Integer setGroupNameFor(String groupName, int groupId, 
-			int groupOwner);
+	Integer setGroupNameFor(String groupName, int groupId, Long groupOwner);
+
 	@Query("SELECT g FROM GroupEntity g WHERE g.groupOwner = ?1 and g.groupName != ?2")
-	List<GroupEntity> findByGroupOwner(int userId, String skipName);	
-	
+	List<GroupEntity> findByGroupOwner(Long userId, String skipName);
+
 }
