+++ hztb-servicemanager/src/main/java/com/svs/hztb/repository/OpinionResponseRepository.java
@@ -9,18 +9,21 @@ import org.springframework.data.repository.query.Param;
 
 import com.svs.hztb.entity.OpinionResponseEntity;
 
-public interface OpinionResponseRepository extends CrudRepository<OpinionResponseEntity, Integer>{
+public interface OpinionResponseRepository extends CrudRepository<OpinionResponseEntity, Integer> {
 	@Query("SELECT o FROM OpinionResponseEntity o WHERE o.responderUserId = :responderUserId and o.responseTime >= :responseTime")
-	List<OpinionResponseEntity> findByResponderUserIdLastUpdatedTime(@Param("responderUserId") int responderUserId, @Param("responseTime") Date responseTime);		
+	List<OpinionResponseEntity> findByResponderUserIdLastUpdatedTime(@Param("responderUserId") Long responderUserId,
+			@Param("responseTime") Date responseTime);
 
-	List<OpinionResponseEntity> findByResponderUserId(int userId);	
-	
-	List<OpinionResponseEntity> findByOpinionId(int opinionId);	
+	List<OpinionResponseEntity> findByResponderUserId(Long userId);
+
+	List<OpinionResponseEntity> findByOpinionId(int opinionId);
+
 	@Query("SELECT o FROM OpinionResponseEntity o WHERE o.opinionId = :opinionId and o.responseTime >= :responseTime")
-	List<OpinionResponseEntity> findByOpinionIdLastUpdatedTime(@Param("opinionId") int opinionId, @Param("responseTime") Date responseTime);		
+	List<OpinionResponseEntity> findByOpinionIdLastUpdatedTime(@Param("opinionId") int opinionId,
+			@Param("responseTime") Date responseTime);
 
 	@Query("SELECT o FROM OpinionResponseEntity o WHERE o.opinionId = :opinionId and o.responderUserId >= :responderUserId")
-	List<OpinionResponseEntity> findByOpinionIdResponderUserId(@Param("opinionId") int opinionId, @Param("responderUserId") int responderUserId);		
+	List<OpinionResponseEntity> findByOpinionIdResponderUserId(@Param("opinionId") int opinionId,
+			@Param("responderUserId") Long responderUserId);
 
-	
 }
