+++ hztb-servicemanager/src/main/java/com/svs/hztb/repository/UserGroupEntityRepository.java
@@ -8,10 +8,10 @@ import org.springframework.data.repository.query.Param;
 
 import com.svs.hztb.entity.UserGroupEntity;
 
-public interface UserGroupEntityRepository extends CrudRepository<UserGroupEntity, Integer>  {
+public interface UserGroupEntityRepository extends CrudRepository<UserGroupEntity, Integer> {
 	@Query("SELECT u FROM UserGroupEntity u WHERE u.id.groupId = :groupId")
 	List<UserGroupEntity> findByGroupId(@Param("groupId") int groupId);
 
 	@Query("SELECT u FROM UserGroupEntity u WHERE u.id.userId = :userId")
-	List<UserGroupEntity> findByUserId(@Param("userId") int userId);
+	List<UserGroupEntity> findByUserId(@Param("userId") Long userId);
 }
